services/timeline-service/requirements.txt
fastapi
uvicorn
kafka-python
redis